Make project more suitable for Code Blocks (CB).
Apparently CB needs this GenArt.cbp file to compile the project. It 
lists things like include files and so on. I will need to keep it in 
sync with our project structure.

In gitignore I added some other useless CB files. I also added obj 
folder to gitignore because it is build result and should not be in 
repository. Also o files are platform dependent.

If this CB thing is to troublesome, then I could use something like 
make.